refactor admin routes into submodules
